fix: 작성 가능한 리뷰 위치 변경 및 유효성 검사 추가
