+++ src/controllers/estimate.controller.ts
@@ -1,24 +1,6 @@
 import { NextFunction, Request, Response } from "express";
 import estimateService from "../services/estimate.service";
 
-// 작성 가능한 리뷰 목록
-async function getWritableEstimates(req: Request, res: Response, next: NextFunction) {
-  try {
-    const clientId = req.auth!.userId;
-
-    const page = Number(req.query.page) || 1;
-    const limit = Number(req.query.limit) || 6;
-
-    const result = await estimateService.getWritableEstimates(clientId, page, limit);
-    res.status(200).json({
-      message: "작성 가능한 리뷰 견적 목록 조회 성공",
-      data: result,
-    });
-  } catch (error) {
-    next(error);
-  }
-}
-
 // 대기 중인 견적 조회
 async function getPendingEstimates(req: Request, res: Response, next: NextFunction) {
   try {
@@ -202,7 +184,6 @@ async function getReceivedEstimates(req: Request, res: Response, next: NextFunct
 }
 
 export default {
-  getWritableEstimates,
   getPendingEstimates,
   sendEstimateToRequest,
   getSentEstimateDetail,
