+++ src/controllers/review.controller.ts
@@ -1,6 +1,12 @@
 import { NextFunction, Request, Response } from "express";
 import reviewService from "../services/review.service";
-import { CreateReviewDto, ReviewIdParamsDto, UpdateReviewDto } from "../dtos/review.dto";
+import {
+  CreateReviewDto,
+  CreateReviewSchema,
+  UpdateReviewDto,
+  UpdateReviewschema,
+} from "../dtos/review.dto";
+import { Review } from "@prisma/client";
 
 // 내가 작성한 리뷰 목록
 async function getMyReviews(req: Request, res: Response, next: NextFunction) {
@@ -24,8 +30,10 @@ async function createReview(
   next: NextFunction,
 ) {
   try {
+    const parsedData = CreateReviewSchema.parse(req.body);
+
     const clientId = req.auth!.userId;
-    const { estimateId, rating, content } = req.body;
+    const { estimateId, rating, content } = parsedData;
     const parsedRating = Number(rating);
     const review = await reviewService.createReview(
       { estimateId, rating: parsedRating, content },
@@ -40,14 +48,16 @@ async function createReview(
 
 // 리뷰 수정
 async function updateReview(
-  req: Request<ReviewIdParamsDto, {}, UpdateReviewDto>,
+  req: Request<{ reviewId: Review["id"] }, {}, UpdateReviewDto>,
   res: Response,
   next: NextFunction,
 ) {
   try {
+    const parsedData = UpdateReviewschema.parse(req.body);
+
     const clientId = req.auth!.userId;
     const { reviewId } = req.params;
-    const { rating, content } = req.body;
+    const { rating, content } = parsedData;
     const updated = await reviewService.updateReview(reviewId, clientId, {
       rating,
       content,
@@ -60,7 +70,11 @@ async function updateReview(
 }
 
 // 리뷰 삭제
-async function deleteReview(req: Request<ReviewIdParamsDto>, res: Response, next: NextFunction) {
+async function deleteReview(
+  req: Request<{ reviewId: Review["id"] }>,
+  res: Response,
+  next: NextFunction,
+) {
   try {
     const clientId = req.auth!.userId;
     const { reviewId } = req.params;
@@ -72,9 +86,28 @@ async function deleteReview(req: Request<ReviewIdParamsDto>, res: Response, next
   }
 }
 
+// 작성 가능한 리뷰 목록
+async function getWritableReviews(req: Request, res: Response, next: NextFunction) {
+  try {
+    const clientId = req.auth!.userId;
+
+    const page = Number(req.query.page) || 1;
+    const limit = Number(req.query.limit) || 6;
+
+    const result = await reviewService.getWritableReviews(clientId, page, limit);
+    res.status(200).json({
+      message: "작성 가능한 리뷰 견적 목록 조회 성공",
+      data: result,
+    });
+  } catch (error) {
+    next(error);
+  }
+}
+
 export default {
   getMyReviews,
   createReview,
   updateReview,
   deleteReview,
+  getWritableReviews,
 };
