+++ src/routers/estimate.router.ts
@@ -1,11 +1,7 @@
 import { Router, RequestHandler } from "express";
 import estimateController from "../controllers/estimate.controller";
-import { verifyAccessToken } from "../middlewares/auth.middleware";
 
 const estimateRouter = Router();
-
-// 작성 가능한 리뷰 목록
-estimateRouter.get("/writable/me", estimateController.getWritableEstimates);
 
 // 대기 중인 견적 조회
 estimateRouter.get("/pending", estimateController.getPendingEstimates as RequestHandler);
