+++ src/services/review.service.ts
@@ -6,9 +6,6 @@ import estimateRepository from "../repositories/estimate.repository";
 
 // 내가 작성한 리뷰 목록
 async function getMyReviews(clientId: Client["id"], page: number = 1, limit: number = 6) {
-  if (!clientId) {
-    throw new BadRequestError("clientId가 필요합니다.");
-  }
   const offset = (page - 1) * limit;
   return reviewRepository.findReviewsByClientId(clientId, offset, limit, page);
 }
@@ -16,12 +13,6 @@ async function getMyReviews(clientId: Client["id"], page: number = 1, limit: num
 // 리뷰 작성
 async function createReview(data: CreateReviewDto, clientId: Client["id"]) {
   const { estimateId, rating, content } = data;
-  if (!estimateId || !content || !clientId) {
-    throw new BadRequestError("필수 입력값이 누락되었습니다.");
-  }
-  if (rating < 1 || rating > 5) {
-    throw new BadRequestError("평점은 1~5 사이여야 합니다.");
-  }
 
   // 견적에서 moverId 조회
   const estimate = await estimateRepository.getEstimateMoverId(estimateId);
@@ -46,7 +37,6 @@ async function updateReview(
   data: Partial<{ rating: Review["rating"]; content: Review["content"] }>,
 ) {
   if (!reviewId) throw new BadRequestError("reviewId가 필요합니다.");
-  if (!data.rating && !data.content) throw new BadRequestError("수정할 내용이 없습니다.");
 
   const review = await reviewRepository.findReviewById(reviewId);
   if (!review) throw new NotFoundError("리뷰를 찾을 수 없습니다.");
@@ -66,9 +56,19 @@ async function deleteReview(reviewId: Review["id"], clientId: Client["id"]) {
   await reviewRepository.deleteReview(reviewId);
 }
 
+// 작성 가능한 리뷰 목록
+async function getWritableReviews(clientId: Client["id"], page: number, limit: number) {
+  if (!clientId) {
+    throw new BadRequestError("clientId가 필요합니다.");
+  }
+  const offset = (page - 1) * limit;
+  return estimateRepository.findWritableEstimatesByClientId(clientId, offset, limit, page);
+}
+
 export default {
   getMyReviews,
   createReview,
   updateReview,
   deleteReview,
+  getWritableReviews,
 };
